+++ DSAII/project2/math-model.hpp
@@ -13,18 +13,18 @@ class MathModel {
 
         // Helper functions for computing statistics
         int computeFactorial(int value);
-        float calcIdleTime(float lambda, float mu, int numServiceChannels);
-        float calcAvgNumPeople(float lambda, float mu, int numServiceChannels);
+        float calcIdleTime(float lambda, float mu, float numServiceChannels);
+        float calcAvgNumPeople(float lambda, float mu, float numServiceChannels);
         float calcAvgTotalTime(float lambda);
         float calcAvgQueueNum(float lambda, float mu);
         float calcAvgTimeInQueue(float lambda);
-        float calcUtilization(float lambda, float mu, int numServiceChannels);
+        float calcUtilization(float lambda, float mu, float numServiceChannels);
     public:
         // Default constructor
         MathModel();
 
         // Parameterized constructor
-        MathModel(float lambda, float mu, int numServiceChannels);
+        MathModel(float lambda, float mu, float numServiceChannels);
 
         // Getters for precalculated statistical values
         float getIdleTime();
@@ -38,7 +38,7 @@ class MathModel {
         float getNextRandomInterval(float avg);
 
         // Runs the statistical functions with the current variables
-        void calculateNewValues(float lambda, float mu, int numServiceChannels);
+        void calculateNewValues(float lambda, float mu, float numServiceChannels);
 };
 
 #endif
